restructured directories based on suggested best practices (see: https://github.com/ixartz/Next-js-Boilerplate)
