+++ app/portfolio/react-router/page.tsx
@@ -1,5 +1,5 @@
 // @ts-ignore
-import PortfolioPageHeader from "@/app/components/blocks/Portfolio/PortfolioPageHeader.tsx";
+import PortfolioPageHeader from "@/components/blocks/Portfolio/PortfolioPageHeader.tsx";
 import "./Main.jsx";
 
 export default function Page() {
