+++ app/sections/About.tsx
@@ -1,4 +1,4 @@
-import Timeline from "../components/blocks/Timeline/Timeline";
+import Timeline from "@/components/blocks/Timeline/Timeline";
 
 export function About() {
   return (
